+++ extras/src/com/mopub/mobileads/FacebookBanner.java
@@ -16,7 +16,7 @@ import com.mopub.common.util.Views;
 import java.util.Map;
 
 /**
- * Tested with Facebook SDK 3.23.1.
+ * Tested with Facebook SDK 4.8.1.
  */
 public class FacebookBanner extends CustomEventBanner implements AdListener {
     public static final String PLACEMENT_ID_KEY = "placement_id";